Revert "tests: add a wrapper for stories and unit test for needed providers"
This reverts commit 3ebe1639a14a6976e34cca3a4a591582ee9c9f39.
